% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiddler.R
\name{get_tiddler}
\alias{get_tiddler}
\title{Get a tiddler}
\usage{
get_tiddler(title, recipe = TW_OPTIONS("recipe"))
}
\arguments{
\item{title}{title of the tiddler to retrieve}

\item{recipe}{string defining which recipe to read from (optional, defaults to "default")}
}
\value{
tiddler information in JSON format
}
\description{
Get a tiddler
}
\examples{
\dontrun{
get_tiddler("GettingStarted")
}
}
