% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tw_html}
\alias{print.tw_html}
\title{Print tw_html object}
\usage{
\method{print}{tw_html}(x, ...)
}
\arguments{
\item{x}{a `tw_html` object.}

\item{...}{additional arguments passed to or from other methods.}
}
\value{
no return
}
\description{
Print tw_html object
}
