% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiddler.R
\name{remove_fields}
\alias{remove_fields}
\title{Remove fields from tiddlers}
\usage{
remove_fields(title, fields, recipe = TW_OPTIONS("recipe"))
}
\arguments{
\item{title}{tiddler title}

\item{fields}{fields to remove}

\item{recipe}{string defining which recipe to write to (optional, defaults to "default")}
}
\value{
no return value
}
\description{
Remove fields from tiddlers
}
