% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{request}
\alias{request}
\title{Perform a request to TiddlyWiki WebServer}
\usage{
request(method, path = "/", query = list(), ...)
}
\arguments{
\item{method}{The method in the httr package, e.g. GET, POST}

\item{path}{The path of request}

\item{query}{The query of request}

\item{...}{Other arguments of request}
}
\value{
The contents of response
}
\description{
Perform a request to TiddlyWiki WebServer
}
