% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tiddler_json2}
\alias{tiddler_json2}
\title{Generate tiddler in json format}
\usage{
tiddler_json2(tiddler)
}
\arguments{
\item{tiddler}{A list for new tiddler}
}
\value{
New tiddler in json format
}
\description{
Generate tiddler in json format
}
