% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{tw_options}
\alias{tw_options}
\title{Set or get options for my package}
\usage{
tw_options(...)
}
\arguments{
\item{...}{Option names to retrieve option values or \code{[key]=[value]} pairs to set options.}
}
\value{
the default and modified options.
}
\description{
Set or get options for my package
}
\section{Supported options}{

The following options are supported
 host: Host of tiddlywiki
 recipe: Recipes are named lists of bags, ordered from lowest priority to highest
 bag: Bags have access controls that determines which users can read or write to them
}

\examples{
tw_options(host = "http://127.0.0.1:8080/")
}
