% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikitext.R
\name{wikitext_backticks_return}
\alias{wikitext_backticks_return}
\alias{wikitext_backticks}
\title{backticks}
\usage{
wikitext_backticks_return(x)
}
\arguments{
\item{x}{character vector}
}
\value{
By default this function outputs (see: \code{cat}) the result. Call the function ending in \code{.return} to catch the result instead.
}
\description{
backticks for wikitext
}
\examples{
wikitext_backticks('FOO')
wikitext_backticks_return('FOO')
}
