% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiddler.R
\name{delete_tiddler}
\alias{delete_tiddler}
\title{Delete a tiddler}
\usage{
delete_tiddler(title)
}
\arguments{
\item{title}{title of the tiddler to retrieve}
}
\value{
no return values
}
\description{
Delete a tiddler
}
\examples{
\dontrun{
delete_tiddler("GettingStarted")
}
}
