% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiddler.R
\name{get_rendered_tiddler}
\alias{get_rendered_tiddler}
\title{Get a rendered tiddler}
\usage{
get_rendered_tiddler(title)
}
\arguments{
\item{title}{title of the tiddler to retrieve}
}
\value{
tiddler rendering in html
}
\description{
Get a rendered tiddler
}
