% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiddler.R
\name{get_tiddler}
\alias{get_tiddler}
\title{Get a tiddler}
\usage{
get_tiddler(title)
}
\arguments{
\item{title}{title of the tiddler to retrieve}
}
\value{
tiddler information in JSON format
}
\description{
Get a tiddler
}
\examples{
if (FALSE) {
get_tiddler("GettingStarted")
}
}
