% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{kable_html}
\alias{kable_html}
\title{Convert Data Frame to HTML Table using kable}
\usage{
kable_html(df, ...)
}
\arguments{
\item{df}{A data frame to be converted to an HTML table.}

\item{...}{Other arguments to be passed to `knitr::kable`.}
}
\value{
A htmltools object containing the HTML representation of the table.
}
\description{
Convert Data Frame to HTML Table using kable
}
\examples{
kable_html(cars[1:10,])
}
