% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiddler.R
\name{put_tiddler}
\alias{put_tiddler}
\title{Put a tiddler}
\usage{
put_tiddler(
  title,
  text,
  type = c("text/vnd.tiddlywiki", "text/x-tiddlywiki", "text/x-markdown", "text/html",
    "text/plain"),
  tags = NULL
)
}
\arguments{
\item{title}{tiddler title}

\item{text}{tiddler text}

\item{type}{tiddler type}

\item{tags}{tiddler tags}
}
\value{
null if success
}
\description{
Put a tiddler
}
\examples{
if (FALSE) {
title <- "New tiddler"
text <- c("!! Section",
          "This is a new tiddler")
type <- "text/vnd.tiddlywiki"
tags <- c("Tag1", "Tag 2")
put_tiddler(title = title, text = text, type = type, tags = tags)
}
}
