% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiddler.R
\name{put_tiddler}
\alias{put_tiddler}
\title{Put a tiddler}
\usage{
put_tiddler(title, text, tags = NULL)
}
\arguments{
\item{title}{tiddler title}

\item{text}{contents of tiddler}

\item{tags}{tags of tiddler}
}
\value{
null if success
}
\description{
Put a tiddler
}
