% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiddler.R
\name{put_tiddler}
\alias{put_tiddler}
\title{Put a tiddler}
\usage{
put_tiddler(
  title,
  text,
  type = c("text/vnd.tiddlywiki", "text/x-tiddlywiki", "text/x-markdown", "text/html",
    "text/plain", "application/json"),
  tags = NULL,
  fields = NULL,
  recipe = TW_OPTIONS("recipe")
)
}
\arguments{
\item{title}{tiddler title}

\item{text}{tiddler text}

\item{type}{tiddler type}

\item{tags}{tiddler tags which is merged with existing tags}

\item{fields}{a named vector for tiddler fields which is merged with existing tags}

\item{recipe}{string defining which recipe to write to (optional, defaults to "default")}
}
\value{
null if success
}
\description{
Put a tiddler
}
\examples{
\dontrun{
title <- "New tiddler"
text <- c("!! Section",
          "This is a new tiddler")
type <- "text/vnd.tiddlywiki"
tags <- c("Tag1", "Tag 2")
fields <- c("F1" = "V1", "F2" = "V2")
put_tiddler(title = title,
            text = text,
            type = type,
            tags = tags,
            fields = fields)
}
}
