% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_base64}
\alias{save_base64}
\title{Save ggplot into base64}
\usage{
save_base64(plot, width = 5, height = 4, dpi = 300, ...)
}
\arguments{
\item{plot}{object for ggplot2 or a function for plot}

\item{width}{image width}

\item{height}{image height}

\item{dpi}{image resolution}

\item{...}{Other arguments for plot function}
}
\value{
character string for base64 image
}
\description{
Save ggplot into base64
}
\examples{

library(ggplot2)
p <- cars |>
    ggplot() +
    geom_point(aes(speed, dist))
p |> save_base64()
}
