% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_field}
\alias{split_field}
\title{Split tiddlywiki field into values}
\usage{
split_field(s)
}
\arguments{
\item{s}{a string}
}
\value{
an vector of values
}
\description{
Split tiddlywiki field into values
}
