% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{tiddler_document}
\alias{tiddler_document}
\title{Format for converting from R Markdown to another tiddler markdown}
\usage{
tiddler_document(
  host = NULL,
  path = NULL,
  tags = NULL,
  fields = NULL,
  use_bookdown = FALSE,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{host}{the host of tiddlywiki web server}

\item{path}{The folder of tiddlywiki. Temp solution as no PUT file api in tiddlywiki WebServer.}

\item{tags}{tiddler tags}

\item{fields}{a named vector for tiddler fields}

\item{use_bookdown}{logical. Use bookdown to generate markdown file.}

\item{overwrite}{whether to overwrite the existing tiddler.}

\item{...}{Other argument pass to md_document}
}
\value{
R Markdown output format to pass to render()
}
\description{
Format for converting from R Markdown to another tiddler markdown
}
\examples{
\dontrun{
library(rmarkdown)
render("input.Rmd")
}
}
