% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{tiddler_document}
\alias{tiddler_document}
\title{Format for converting from R Markdown to another tiddler markdown}
\usage{
tiddler_document(path = ".")
}
\arguments{
\item{path}{The folder of tiddlywiki}
}
\value{
R Markdown output format to pass to render()
}
\description{
Format for converting from R Markdown to another tiddler markdown
}
\examples{
if (FALSE) {
library(rmarkdown)
render("input.Rmd", tiddler_document())
}
}
