% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tiddler_json2}
\alias{tiddler_json2}
\title{Generate tiddler in json format}
\usage{
tiddler_json2(tiddler)
}
\arguments{
\item{title}{tiddler title}

\item{text}{tiddler text}

\item{type}{tiddler type}

\item{tags}{a vector for tiddler tags}

\item{fields}{a named vector for tiddler fields.}

\item{format}{export format as json or list}
}
\value{
New tiddler in json format
}
\description{
Generate tiddler in json format
}
