% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{tw_options}
\alias{tw_options}
\title{Set or get options for my package}
\usage{
tw_options(...)
}
\arguments{
\item{...}{Option names to retrieve option values or \code{[key]=[value]} pairs to set options.}
}
\value{
the default and modified options.
}
\description{
Set or get options for my package
}
\section{Supported options}{

The following options are supported
\itemize{
 \item{\code{host}}{\code{character} host of tiddlywiki}
}
}

\examples{
tw_options(host = "http://127.0.0.1:8080/")
}
