% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tw_table}
\alias{tw_table}
\title{Convert data.frame into table of TiddlyWiki}
\usage{
tw_table(df, collapse = "\\n")
}
\arguments{
\item{df}{data.frame object}

\item{collapse}{an optional character string to separate the results.}
}
\value{
character string for table in TiddlyWiki
}
\description{
Convert data.frame into table of TiddlyWiki
}
\examples{
cars |>
    dplyr::slice(1:10) |>
    tw_table()
}
