% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_widget.R
\name{tw_widget}
\alias{tw_widget}
\title{Create a tiddlywiki widget from htmlwidget}
\usage{
tw_widget(widget, is_cat = FALSE)
}
\arguments{
\item{widget}{an object of htmlwidget}

\item{is_cat}{whether to show results on screen}
}
\value{
a new tiddlywiki widget
}
\description{
Create a tiddlywiki widget from htmlwidget
}
\examples{
library(leaflet)
content <- paste(sep = "<br/>",
                 "<b><a href='http://www.samurainoodle.com'>Samurai Noodle</a></b>",
                 "606 5th Ave. S",
                 "Seattle, WA 98138"
)

widget <- leaflet() \%>\% addTiles() \%>\%
    addPopups(-122.327298, 47.597131, content,
                  options = popupOptions(closeButton = FALSE)
    )
tw_widget(widget)
}
